+++ src/terrain_base_tile.hpp
@@ -81,7 +81,7 @@ public:
     void initGL(void);
 
     /* Assuming all the textures are set up, issues the draw call. */
-    void draw(unsigned int instanceCount) const;
+    void draw(size_t instanceCount) const;
 
     /* Tears down the VAO. */
     void teardownGL(void) const;
